early splitting into file structure to stay organized, next step is to add the openCV sdk
